requirements.txt
Flask==0.12.1
GoogleAppEngineCloudStorageClient==1.9.22.1
google-cloud-storage==1.6.0
google-cloud-pubsub==0.29.0
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.9.6
jsonschema==2.6.0
MarkupSafe==1.0
pymongo==3.5.1
python-digits==2.0
redis==2.10.5
Werkzeug==0.12.1
